+++ tests/mleap/test_mleap_model_export.py
@@ -4,6 +4,7 @@ from unittest import mock
 from packaging.version import Version
 
 import numpy as np
+import pandas as pd
 import pyspark
 from pyspark.ml.pipeline import Pipeline
 from pyspark.ml.wrapper import JavaModel
@@ -177,3 +178,14 @@ def test_spark_module_model_save_with_relative_path_and_valid_sample_input_produ
         assert os.path.exists(config_path)
         config = Model.load(config_path)
         assert mlflow.mleap.FLAVOR_NAME in config.flavors
+
+
+@pytest.mark.large
+def test_mleap_module_model_save_with_invalid_sample_input_type_raises_exception(
+    spark_model_iris, model_path
+):
+    with pytest.raises(Exception, match="must be a PySpark dataframe"):
+        invalid_input = pd.DataFrame()
+        mlflow.spark.save_model(
+            spark_model=spark_model_iris.model, path=model_path, sample_input=invalid_input
+        )
